+++ src/Resources/lang/pl/app.php
@@ -6,6 +6,8 @@ return [
             'error' => [
                 'resource-not-found' => 'Nie znaleziono zasobu.',
                 'empty-payload' => 'Nie wysłano żadnych produktów.',
+                'not-a-product-list' => 'Treść żądania musi być listą produktów.',
+                'not-a-product-object' => 'Ten wpis nie jest obiektem produktu.',
             ],
         ],